Move test for period
